refactor: overview 폴더 sections·ai-summary 구조로 정리
